% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignCircle.R
\name{AlignedTrueLenght}
\alias{AlignedTrueLenght}
\title{Aligned sequence true length without gaps}
\usage{
AlignedTrueLenght(seq)
}
\arguments{
\item{seq}{vector of chars}
}
\value{
length
}
\description{
Aligned sequence true length without gaps
}
