% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment2Table.R
\name{alignment2Table}
\alias{alignment2Table}
\title{Sequence Alignment to Table}
\usage{
alignment2Table(SeqInfo, SeqAligned)
}
\arguments{
\item{SeqInfo}{table of sequence information generated using \code{\link{getSeqInfo}}}

\item{SeqAligned}{list of sequnece vectors \code{\link{alignment2Fasta}}}
}
\value{
A data.frame object
}
\description{
Converts sequence alignment to data frame format
}
