% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentNoGaps.R
\name{alignmentNoGaps}
\alias{alignmentNoGaps}
\title{Get sequence alignment list of sequences without gaps}
\usage{
alignmentNoGaps(SeqAligned)
}
\arguments{
\item{SeqAligned}{list of sequence vectors \code{\link{alignment2Fasta}}}
}
\value{
A list of sequences without gaps
}
\description{
In Aligned seq out location of non gap sequence (list)
it used in the algorithm of drawing circle sequences
}
