% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentNoGapsLinks.R
\name{alignmentNoGapsLinks}
\alias{alignmentNoGapsLinks}
\title{Get links for alignment regions}
\usage{
alignmentNoGapsLinks(SeqAligned, SeqHighLights)
}
\arguments{
\item{SeqAligned}{list of sequence vectors \code{\link{alignment2Fasta}}}

\item{SeqHighLights}{list of sequence vectors \code{\link{alignmentNoGaps}}}
}
\value{
a list for the location of sequence alignment links between sequences for regions
similar in sequences
}
\description{
In aligned seq out location of non gaped sequence in the original sequence (list)
used for drawing sequences in circos
}
