% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentStats.R
\name{AlignmentStatsPerSeq}
\alias{AlignmentStatsPerSeq}
\title{Sequence alignment statistics table per sequence}
\usage{
AlignmentStatsPerSeq(SeqInfo, SeqAligned)
}
\arguments{
\item{SeqInfo}{table of sequence information generated using \code{\link{getSeqInfo}}}

\item{SeqAligned}{list of aligned vectors contains aligned sequences using \code{\link{alignment2Fasta}}}
}
\value{
Table
}
\description{
Sequence alignment statistics table per sequence
}
