% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignCircle.R
\name{AlignSplit}
\alias{AlignSplit}
\title{Split sequence alignment object}
\usage{
AlignSplit(seq)
}
\arguments{
\item{seq}{a single aligned sequence, as a character string}
}
\value{
character vector of the sequence split into individual bases
}
\description{
Split sequence alignment object
}
