% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignstatplot.R
\name{alignstatplot}
\alias{alignstatplot}
\title{Perform all analysis as a pipeline}
\usage{
alignstatplot(
  SeqFile,
  AlignMethod = "ClustalW",
  AnnoFile = "",
  OutFolder = "output",
  MaxMissPer = 0.2,
  MaxCluster = 4,
  MinimumClusterLength = 3,
  Verbose = T,
  fontscale = 0
)
}
\arguments{
\item{SeqFile}{Fasta File path}

\item{AlignMethod}{(optional) Alignment tool that will be used. "ClustalW", "ClustalOmega", and "Muscle" are supported}

\item{AnnoFile}{Annotation file}

\item{OutFolder}{(optional) Output folder name}

\item{MaxMissPer}{(optional) Maximum Missing in SNPs to remove}

\item{MaxCluster}{(optional) Maximum number of SNP cluster to report}

\item{MinimumClusterLength}{(optional) Minimum Nucleotides in one cluster to be reported}

\item{Verbose}{(optional) verbose}
}
\value{
plots
}
\description{
Perform all analysis as a pipeline
}
