% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeWithGenes.R
\name{AnnotationTable}
\alias{AnnotationTable}
\title{Read Annotation file}
\usage{
AnnotationTable(AnnoFilePath, GeneInfoTable)
}
\arguments{
\item{AnnoFilePath}{Annotation file path}

\item{GeneInfoTable}{Gene information Table}
}
\value{
Table
}
\description{
Read Annotation file
}
