% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawSeqLogo.R
\name{drawSeqLogo}
\alias{drawSeqLogo}
\title{Draw the aligned sequences in logo format}
\usage{
drawSeqLogo(
  SeqAligned,
  SplitLen = 50,
  joinPlot = T,
  outFolder = "",
  width = 15,
  height = 30
)
}
\arguments{
\item{SeqAligned}{Sequences aligned \code{\link{alignment2Fasta}}}

\item{SplitLen}{Integer -- number of nucleotides in every plot}

\item{joinPlot}{True/False Plot to one plot or to several plots}

\item{outFolder}{Specify the folder for plots}

\item{width}{Plot height}

\item{height}{Plot width}
}
\value{
plot
}
\description{
Drawing of the aligned sequences in logo format showing the percentage for every nucleotide
}
