% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathSlash.R
\name{formatFolderPath}
\alias{formatFolderPath}
\title{Format Folder Path according to operating system}
\usage{
formatFolderPath(Folder)
}
\arguments{
\item{Folder}{Folder Path}
}
\value{
Slash
}
\description{
Format Folder Path according to operating system
}
