% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqLocateCluster.R, R/plotTreeWithGenes.R
\name{GeneInfoForClusterPlot}
\alias{GeneInfoForClusterPlot}
\title{Return Gene information table for clusterplot}
\usage{
GeneInfoForClusterPlot(SeqInfo)

GeneInfoForClusterPlot(SeqInfo)
}
\arguments{
\item{SeqInfo}{Information of the sequences \code{\link{getSeqInfo}}}
}
\value{
data.frame

data.frame
}
\description{
Return Gene information table for clusterplot

Return Gene information table for clusterplot
}
