% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiallelicByFreq.R
\name{getBiallelicByFreq}
\alias{getBiallelicByFreq}
\title{Return biallelic nucleotide location in frequency table}
\usage{
getBiallelicByFreq(NucCount)
}
\arguments{
\item{NucCount}{Table of nucleotide frequency \code{\link{nucFrequency}}}
}
\value{
a vector of nucleotide locations in consensus sequence
}
\description{
takes the nucleotide frequency and returns the location of
biallelic loci such as A/G
}
