% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqLocateCluster.R
\name{getClusterAnnotation}
\alias{getClusterAnnotation}
\title{Get Cluster annotation table for SNP groups in sequences}
\usage{
getClusterAnnotation(
  SeqInfo,
  SeqWithClusters,
  ClusterMax = 3,
  MinimumClusterLength = 3
)
}
\arguments{
\item{SeqInfo}{Information of the sequences \code{\link{getSeqInfo}}}

\item{SeqWithClusters}{Sequences with clusters generated by \code{\link{SeqLocateCluster}}}

\item{ClusterMax}{Max number of clusters in the sequences}

\item{MinimumClusterLength}{Minimum length of clustered nucleotides to consider as a group cluster}
}
\value{
Table
}
\description{
Get Cluster annotation table for SNP groups in sequences
}
