% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterTable.R
\name{getClusterTable}
\alias{getClusterTable}
\title{Get Cluster Table of SNPs}
\usage{
getClusterTable(Cluster)
}
\arguments{
\item{Cluster}{Cluster object generated using \code{\link{SNPCluster}}}
}
\value{
data.frame
}
\description{
Get Cluster Table of SNPs
}
