% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistanceMatrixTabel.R
\name{getDistanceMatrixTabel}
\alias{getDistanceMatrixTabel}
\title{Get distance matrix for sequence alignment}
\usage{
getDistanceMatrixTabel(seqInfo, seqAlignment)
}
\arguments{
\item{seqInfo}{Table of sequence information \code{\link{getSeqInfo}}}

\item{seqAlignment}{sequence alignment list generated using \code{\link[msa]{msa}}}
}
\value{
matrix of sequences similarity distances
}
\description{
Get distance matrix for sequence alignment
}
