% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefGenotypeForbiallelic.R
\name{getRefGenotypeForbiallelic}
\alias{getRefGenotypeForbiallelic}
\title{Get a reference sequence using biallelic nucleotides}
\usage{
getRefGenotypeForbiallelic(NucCount, biallelic)
}
\arguments{
\item{NucCount}{Table of nucleotide frequency \code{\link{nucFrequency}}}

\item{biallelic}{location of biallelic nucleotides \code{\link{getBiallelicByFreq}}}
}
\value{
a vector of nucleotides can be used as a reference
}
\description{
#' Get a sequence can be used as a reference for sequence binary
conversion using for only biallelic nucleotides
}
