% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSectorWidth.R
\name{getSectorWidth}
\alias{getSectorWidth}
\title{Calculate Zooming factor and sector width for circular plot with links}
\usage{
getSectorWidth(SeqInfo, ConsLength, ZoomFactor)
}
\arguments{
\item{SeqInfo}{Table of sequence information \code{\link{getSeqInfo}}}

\item{ConsLength}{Length of the sequence in base pair}

\item{ZoomFactor}{A number between 1-10 for the size of the consensus representation
in relative to the other sequences}
}
\value{
A table for every sector name (sequence) and zooming ratio
}
\description{
Calculate Zooming factor and sector width for circular plot with links
}
