% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqColors.R
\name{getSeqColors}
\alias{getSeqColors}
\title{get sequences colors}
\usage{
getSeqColors(ColorsN)
}
\arguments{
\item{ColorsN}{number of colors}
}
\value{
character
}
\description{
get sequences colors
}
\examples{
getSectorColors(10)
}
