% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqInfo.R
\name{getSeqInfo}
\alias{getSeqInfo}
\title{Extract sequence information from a fasta file}
\usage{
getSeqInfo(SeqFilePath)
}
\arguments{
\item{SeqFilePath}{Fasta file path}
}
\value{
A table of sequence information
\itemize{
  \item Name - Sequence Name
  \item Length - Sequence in base pairs
}
}
\description{
Extract sequence information from a fasta file
}
\examples{
SeqFile<-system.file("extdata","Example_Small.fasta",package = "alignstatplot")
SeqInfo<-getSeqInfo(SeqFile)
SeqInfo
}
