% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqLogo.R
\name{getSeqLogo}
\alias{getSeqLogo}
\title{Split aligned sequence to list of vectors}
\usage{
getSeqLogo(SeqAligned, splitRange = 20)
}
\arguments{
\item{SeqAligned}{Sequences aligned \code{\link{alignment2Fasta}}}

\item{splitRange}{Integer -- number of nucleotides in every plot}
}
\value{
A list of vectors
}
\description{
Split aligned sequences to list of vectors where every vector
contains a list of a part (split range) of every aligned sequence
}
