% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentNoGapsLinks.R
\name{getSeqRegion}
\alias{getSeqRegion}
\title{Get regions of sequences in sequence aligned with marks}
\usage{
getSeqRegion(seqMarked)
}
\arguments{
\item{seqMarked}{character vector of aligned sequence}
}
\value{
list of locations
}
\description{
Get regions of sequences in sequence aligned with marks
}
