% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentNoGapsLinks.R
\name{getSeqRegionCI}
\alias{getSeqRegionCI}
\title{getSeqRegion_ C implementation}
\usage{
getSeqRegionCI(seqMarked)
}
\arguments{
\item{seqMarked}{character vector of aligned sequence}
}
\value{
list of vectors of regions locations
}
\description{
getSeqRegion_ C implementation
}
