% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTree.R
\name{getTree}
\alias{getTree}
\title{Get Tree from a distance Matrix}
\usage{
getTree(DistMatrixTable)
}
\arguments{
\item{DistMatrixTable}{A distance matrix for similarity matrix across the genes
generated using \code{\link{getDistanceMatrixTabel}}}
}
\value{
an object of class "phylo"
}
\description{
Get Tree from a distance Matrix
}
