% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucFrequency.R
\name{nucFrequency}
\alias{nucFrequency}
\title{Calculates nucleotide Frequency across sequences}
\usage{
nucFrequency(TargetTable)
}
\arguments{
\item{TargetTable}{A Table of aligned sequences (rows) x nucleotides (columns)
generated using \code{\link{alignment2Table}}}
}
\value{
Table of nucleotide frequencies across the sequences including
the missing nucleotide (N)
}
\description{
Calculates nucleotide Frequency across sequences
}
