% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucTableFilter.R
\name{nucTableFilter}
\alias{nucTableFilter}
\title{Filter aligned seq table from missing and mono}
\usage{
nucTableFilter(SeqTable, MaxMissPer = 0.2, removeMono = T)
}
\arguments{
\item{SeqTable}{ist of sequence vectors \code{\link{alignment2Fasta}}}

\item{MaxMissPer}{Maximum percentage of missing nucleotides}

\item{removeMono}{bool value remove monomorphic variations}
}
\value{
data frame
}
\description{
Filter aligned seq table from missing and mono
}
