% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucTableFreqHeatmapSplit.R
\name{nucTableFreqHeatmapSplit}
\alias{nucTableFreqHeatmapSplit}
\title{Generate several plots for sequence alignment}
\usage{
nucTableFreqHeatmapSplit(
  SeqAlignedTable,
  maxN = 100,
  cex.NucLabels = 7,
  cex.SeqLabels = 10,
  ...
)
}
\arguments{
\item{SeqAlignedTable}{list of aligned vectors contains aligned sequences using \code{\link{alignment2Fasta}}}

\item{maxN}{Maximum number of nucleotides in every plot}

\item{cex.NucLabels}{Label font size}

\item{cex.SeqLabels}{Sequence name font size}

\item{...}{additional arguments forwarded to \code{\link{nucTableFreqHeatmap}}
(e.g. \code{colors}, \code{heights}, \code{title})}
}
\value{
list of plots
}
\description{
Generate several plots for sequence alignment
}
