% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentStats.R
\name{percentFormat}
\alias{percentFormat}
\title{Formatting numbers as a percentage}
\usage{
percentFormat(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{value}

\item{digits}{number of digits}

\item{format}{formatting}

\item{...}{...}
}
\value{
Percent ion format
}
\description{
Formatting numbers as a percentage
}
