% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignCircle.R
\name{plotAlignCircle}
\alias{plotAlignCircle}
\title{Plot Sequence Alignment as circle plot}
\usage{
plotAlignCircle(path, SeqFormat = "fasta", SeqFontSize = 1, colors = NULL)
}
\arguments{
\item{path}{Sequence alignment Fasta path}

\item{SeqFormat}{fasta clustalw}

\item{SeqFontSize}{sequence labels font size}

\item{colors}{sector colors, one per sequence (default: \code{\link{getSeqColors}})}
}
\value{
plot
}
\description{
Plot Sequence Alignment as circle plot
}
