% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIdentityDistribution.R
\name{plotIdentityDistribution}
\alias{plotIdentityDistribution}
\title{Plot the distribution of pairwise sequence identity}
\usage{
plotIdentityDistribution(DistMatrixTable, colors = NULL, bins = 20)
}
\arguments{
\item{DistMatrixTable}{A distance matrix generated using \code{\link{getDistanceMatrixTabel}}}

\item{colors}{fill/line color, length 1 or 2 (default: the package's sequential blue ramp)}

\item{bins}{number of histogram bins}
}
\value{
a \code{ggplot} object
}
\description{
Plot the distribution of pairwise sequence identity
}
