% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeWithGenes.R
\name{plotTreeWithGenes}
\alias{plotTreeWithGenes}
\title{Title}
\usage{
plotTreeWithGenes(
  SeqInfo,
  Alignment,
  AnnoFilePath,
  geneArrowFill = "white",
  rel_widths = c(1, 2),
  label_size = 1
)
}
\arguments{
\item{SeqInfo}{table of sequence information generated using \code{\link{getSeqInfo}}}

\item{Alignment}{sequence alignment object}

\item{AnnoFilePath}{Annotation file path}

\item{geneArrowFill}{fill color for the gene arrow bodies}

\item{rel_widths}{relative widths of the tree panel and gene-structure panel}

\item{label_size}{panel label font size, forwarded to \code{cowplot::plot_grid}}
}
\value{
plot
}
\description{
Title
}
