% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeq.R
\name{readSeq}
\alias{readSeq}
\title{Read and reformat fasta sequence}
\usage{
readSeq(fastaPath, outPath)
}
\arguments{
\item{fastaPath}{path of fasta file}

\item{outPath}{output folder path}
}
\value{
return a list of vector of chars. Each element is a sequence
object of the class 'SeqFastadna'
}
\description{
Read and reformat fasta sequence
}
