% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqAlign.R
\name{seqAlign}
\alias{seqAlign}
\title{Perfom sequence alignment}
\usage{
seqAlign(SeqFile, AlignMethod, SeqType = "dna", order = "input")
}
\arguments{
\item{SeqFile}{sequence file path}

\item{AlignMethod}{specifies the multiple sequence alignment to be used; currently, "ClustalW",
"ClustalOmega", and "Muscle" are supported.}

\item{SeqType}{"dna"}

\item{order}{Ordering sequences in the list random or according to the input}
}
\value{
Returns MsaDNAMultipleAlignment \code{\link[msa]{msa}}
}
\description{
Perfom sequence alignment
}
