% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqLocateCluster.R
\name{SeqLocateCluster}
\alias{SeqLocateCluster}
\title{Locate Clustered SNPs in sequence}
\usage{
SeqLocateCluster(SeqAligned, ClusterTable)
}
\arguments{
\item{SeqAligned}{list of sequence vectors \code{\link{alignment2Fasta}}}

\item{ClusterTable}{Table generated using getClusterTable function \code{\link{getClusterTable}}}
}
\value{
List of sequences , where clustered SNPs are marked by number
}
\description{
Locate Clustered SNPs in sequence
}
