% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqRefCommon.R
\name{seqRefCommon}
\alias{seqRefCommon}
\title{Create reference form a batch/single sequence}
\usage{
seqRefCommon(SeqAlignedTable, RefsNames)
}
\arguments{
\item{SeqAlignedTable}{list of aligned vectors contains aligned sequences using \code{\link{alignment2Fasta}}}

\item{RefsNames}{vector of sequences names chosen as reference}
}
\value{
vector of single sequence can be used as a reference
}
\description{
Create reference form a batch/single sequence
}
