% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTableToBinary.R
\name{seqTableToBinary}
\alias{seqTableToBinary}
\title{Convert table of aligned sequence to binary format}
\usage{
seqTableToBinary(SeqAlignedTable, Refs, RemoveNonRefNuc = T, RefsNames = T)
}
\arguments{
\item{SeqAlignedTable}{list of aligned vectors contains aligned sequences using \code{\link{alignment2Fasta}}}

\item{Refs}{vector of sequences IDs to be used as a reference for
bionary conversion}

\item{RemoveNonRefNuc}{logical -- remove nucleotide absent in reference (Insertions)}

\item{RefsNames}{logical is the provided is reference IDs or a reference as avector generated
using \code{\link{seqRefCommon}} or \code{\link{getRefGenotypeForbiallelic}}}
}
\value{
Table of 0/1 variations
}
\description{
Convert Sequence Table aligned to Binary Format according
to reference
}
