% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqWithLast.R
\name{seqWithLast}
\alias{seqWithLast}
\title{Sequence of numbers including  the end number if it was not reached}
\usage{
seqWithLast(from, to, by)
}
\arguments{
\item{from}{start}

\item{to}{end}

\item{by}{step}
}
\value{
vector
}
\description{
#return a sequence of numbers including last number if
it was not included in the sequence
}
