% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPCluster.R
\name{SNPCluster}
\alias{SNPCluster}
\title{Performs MCA or PCA clustering depending on data type}
\usage{
SNPCluster(SeqAlignedTable, ncp = 20)
}
\arguments{
\item{SeqAlignedTable}{list of aligned vectors contains aligned sequences using \code{\link{alignment2Fasta}}}

\item{ncp}{number of dimensions kept in the results (by default 5)}
}
\value{
an HCPC object  \code{\link[FactoMineR]{HCPC}}
}
\description{
The function perfoms clustering analysis dependiong on the data format
if the provided data is binary based {0, 1} it will use PCA analysis a PCA list components please see \code{\link[FactoMineR]{PCA}}
, while if allele-based {A,C,T,G} it will use \code{\link[FactoMineR]{MCA}}
}
