% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqLocateCluster.R
\name{SNPClusterPlot}
\alias{SNPClusterPlot}
\title{Plot SNP cluster locations on genes}
\usage{
SNPClusterPlot(SeqInfo, SeqAligned, Cluster, MaxCluster, MinimumClusterLength)
}
\arguments{
\item{SeqInfo}{table of sequence information generated using \code{\link{getSeqInfo}}}

\item{SeqAligned}{list of aligned vectors contains aligned sequences using \code{\link{alignment2Fasta}}}

\item{Cluster}{object of HCPC \code{\link{SNPCluster}}}

\item{MaxCluster}{Maximum number of clusters to draw default 3}

\item{MinimumClusterLength}{Minimum number of base pairs for a cluster to be consider 3bp.}
}
\value{
plot
}
\description{
Plot SNP cluster locations on genes
}
