% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqLocateCluster.R
\name{SNPClusterPlotWithTree}
\alias{SNPClusterPlotWithTree}
\title{Plot SNP cluster on genes along with Phylogenetic tree}
\usage{
SNPClusterPlotWithTree(
  SeqInfo,
  Alignment,
  Cluster,
  MaxCluster = 3,
  MinimumClusterLength = 3,
  geneArrowFill = "white",
  rel_widths = c(1, 2),
  label_size = 1
)
}
\arguments{
\item{SeqInfo}{table of sequence information generated using \code{\link{getSeqInfo}}}

\item{Alignment}{sequence alignment object}

\item{Cluster}{object of HCPC \code{\link{SNPCluster}}}

\item{MaxCluster}{Maximum number of clusters to draw default 3}

\item{MinimumClusterLength}{Minimum number of base pairs for a cluster to be consider 3bp.}

\item{geneArrowFill}{fill color for the gene arrow bodies}

\item{rel_widths}{relative widths of the tree panel and gene-structure panel}

\item{label_size}{panel label font size, forwarded to \code{cowplot::plot_grid}}
}
\value{
plot
}
\description{
Plot SNP cluster on genes along with Phylogenetic tree
}
