% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment2Fasta.R
\name{alignment2Fasta}
\alias{alignment2Fasta}
\title{Write Sequence Alignment to fasta or return it as a list of sequences}
\usage{
alignment2Fasta(alignment, SeqInfo, filename = "", format = "blocks")
}
\arguments{
\item{alignment}{sequence alignment object \code{\link[ape]{DNAbin}} \code{\link{seqAlign}}}

\item{SeqInfo}{Information of the sequences \code{\link{getSeqInfo}}}

\item{filename}{If it was set, the file name sequences will be written in file}

\item{format}{if "blocks" (default), also write the alignment to \code{filename} via \code{\link[ape]{write.dna}}}
}
\value{
A list of vectors of aligned sequences or write it to file
}
\description{
Write Sequence Alignment to fasta or return it as a list of sequences
}
\examples{
\dontrun{
# requires the ClustalW executable to be installed and on the PATH
seqFile<-system.file("extdata","Example_Small.fasta",package = "alignstatplot")
SeqInfo<-getSeqInfo(seqFile)
myClustalWAlignment <- seqAlign(seqFile,"ClustalW")
SeqAligned<-alignment2Fasta(myClustalWAlignment,SeqInfo)
SeqAligned
}
}
