% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment2Fasta.R
\name{alignment2Fasta}
\alias{alignment2Fasta}
\title{Write Sequence Alignment to fasta or return it as a list of sequences}
\usage{
alignment2Fasta(alignment, SeqInfo, filename = "")
}
\arguments{
\item{alignment}{sequence alignment object \code{\link[msa]{msa}} \code{\link{seqAlign}}}

\item{SeqInfo}{Information of the sequences \code{\link{getSeqInfo}}}

\item{filename}{If it was set, the file name sequences will be written in file}
}
\value{
A list of vectors of aligned sequences or write it to file
}
\description{
Write Sequence Alignment to fasta or return it as a list of sequences
}
\examples{
seqFile<-system.file("extdata","sequence_few.fasta",package = "alignstatplot")
myClustalWAlignment <- seqAlign(seqFile,"ClustalW")
SeqAligned<-alignment2Fasta(myClustalWAlignment,SeqInfo)
SeqAligned
}
