% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceHeatmap.R
\name{distanceHeatmap}
\alias{distanceHeatmap}
\title{Plot distance matrix as heatmap}
\usage{
distanceHeatmap(DistTable, fontsizescale = 1.5)
}
\arguments{
\item{DistTable}{Table of distance matrix generated using \code{\link{getDistanceMatrixTabel}}}

\item{fontsizescale}{scale of font size}
}
\value{
heatmap plot
}
\description{
Plot distance matrix as heatmap
}
