% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawConsWithGenes.R
\name{drawConsWithGenes}
\alias{drawConsWithGenes}
\title{Draw circos plot for sequence alignment for genes with the consensus sequence}
\usage{
drawConsWithGenes(SeqInfo, SeqAligned, consZoomFactor = 3, cex.SeqLabels = 0.5)
}
\arguments{
\item{SeqInfo}{Table of sequence information generated using \code{\link{getSeqInfo}}}

\item{SeqAligned}{list of aligned vectors contains aligned sequences using \code{\link{alignment2Fasta}}}

\item{consZoomFactor}{A number between 1-10 for the size of the consensus representation
in relative to the other sequences}

\item{cex.SeqLabels}{A number for sequence labels font size}
}
\value{
plot of sequence alignment for genes with the consensus sequence
}
\description{
Draw circos plot for sequence alignment for genes with the consensus sequence
}
