% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawConsWithNoGenes.R
\name{drawConsWithNoGenes}
\alias{drawConsWithNoGenes}
\title{Draw Consensus Circos without Links}
\usage{
drawConsWithNoGenes(
  SeqInfo,
  SeqAligned,
  cex.SeqLabels = 0.5,
  cex.bpLabels = 0.3
)
}
\arguments{
\item{SeqInfo}{Table of sequence information generated using \code{\link{getSeqInfo}}}

\item{SeqAligned}{list of aligned vectors contains aligned sequences using \code{\link{alignment2Fasta}}}

\item{cex.SeqLabels}{A number for sequence labels font size}

\item{cex.bpLabels}{A number for base pair labels font size}
}
\value{
plot an alignment circle
}
\description{
Draw Consensus Circos without Links
}
