% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawGenes.R
\name{drawGenes}
\alias{drawGenes}
\title{Drawing sequence basic structure including inverting the consensus direction without links}
\usage{
drawGenes(mycircos.Seq.Sectors, inv)
}
\arguments{
\item{mycircos.Seq.Sectors}{Table of genes as sectors of the circle plot}

\item{inv}{The name of the sequence that will be inverted -- mostly the consensus}
}
\value{
draws genes and consensus as a part of the drawing alignment function \code{\link{drawConsWithGenes}}
}
\description{
Drawing sequence basic structure including inverting the consensus direction without links
}
