% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucFrequencyPlot.R
\name{nucFrequencyPlot}
\alias{nucFrequencyPlot}
\title{Plot the frequency of nucleotides across the sequence}
\usage{
nucFrequencyPlot(NucCount, xlabel = TRUE, cex.NucLabels = 7)
}
\arguments{
\item{NucCount}{Table of nueleotide frequency across the aligned sequences \code{\link{nucFrequency}}}

\item{xlabel}{font size of sequences names}

\item{cex.NucLabels}{font size of nucleotides}
}
\value{
A plot for the frequency of nucleotides
}
\description{
Plot the frequency of nucleotides across the sequence
}
