% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucTableFreqHeatmap.R
\name{nucTableFreqHeatmap}
\alias{nucTableFreqHeatmap}
\title{Heatmap and frequency plots for sequence alignment}
\usage{
nucTableFreqHeatmap(SeqAlignedTable, cex.NucLabels = 7, cex.SeqLabels = 10)
}
\arguments{
\item{SeqAlignedTable}{list of aligned vectors contains aligned sequences using \code{\link{alignment2Fasta}}}

\item{cex.NucLabels}{Label font size}

\item{cex.SeqLabels}{Sequence name font size}
}
\value{
Heatmap and frequency plots
}
\description{
Heatmap and frequency plots for sequence alignment
}
