% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucTableHeatmap.R
\name{nucTableHeatmap}
\alias{nucTableHeatmap}
\title{Plot aligned sequence variation as heatmap}
\usage{
nucTableHeatmap(SeqAlignedTable, cex.NucLabels = 7, cex.SeqLabels = 10)
}
\arguments{
\item{SeqAlignedTable}{list of aligned vectors contains aligned sequences using \code{\link{alignment2Fasta}}}

\item{cex.NucLabels}{Label font size}

\item{cex.SeqLabels}{Sequence name font size}
}
\value{
Heatmap plot
}
\description{
Plot aligned sequence variation as heatmap
}
