% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignCircle.R
\name{plotAlignCircle}
\alias{plotAlignCircle}
\title{Plot Sequence Alignment as circle plot}
\usage{
plotAlignCircle(path, SeqFormat = "fasta", SeqFontSize = 1)
}
\arguments{
\item{path}{Sequence alignment Fasta path}

\item{format}{fasta clustalw}
}
\value{
plot
}
\description{
Plot Sequence Alignment as circle plot
}
