% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA for genes / Samples using distance matrix generatd from alignment}
\usage{
plotPCA(seqInfo, seqAlignment, ncluster = 4, labelsize = 10, showlabels = T)
}
\arguments{
\item{seqInfo}{seqInfo table of sequence information generated using \code{\link{getSeqInfo}}}

\item{seqAlignment}{sequence alignment object}

\item{ncluster}{number of cluster default 4}
}
\value{
plot of PCA
}
\description{
Plot PCA for genes / Samples using distance matrix generatd from alignment
}
