% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeWithGenes.R
\name{plotTreeWithGenes}
\alias{plotTreeWithGenes}
\title{Title}
\usage{
plotTreeWithGenes(SeqInfo, Alignment, AnnoFilePath)
}
\arguments{
\item{SeqInfo}{table of sequence information generated using \code{\link{getSeqInfo}}}

\item{Alignment}{sequence alignment object \code{\link[msa]{msa}} \code{\link{seqAlign}}}

\item{AnnoFilePath}{Annotation file path}
}
\value{
plot
}
\description{
Title
}
