% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeWithRuler.R
\name{plotTreeWithRuler}
\alias{plotTreeWithRuler}
\title{Plot phylogenetic tree along with distance ruler}
\usage{
plotTreeWithRuler(SeqInfo, myClustalWAlignment)
}
\arguments{
\item{SeqInfo}{table of sequence information generated using \code{\link{getSeqInfo}}}

\item{myClustalWAlignment}{sequence alignment object}
}
\value{
Tree plot
}
\description{
Plot phylogenetic tree along with distance ruler
}
