% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSeqPlotList.R
\name{saveSeqPlotList}
\alias{saveSeqPlotList}
\title{save plot lists}
\usage{
saveSeqPlotList(PlotList, outFolder, width = 15, height = 10)
}
\arguments{
\item{PlotList}{list of plots}

\item{outFolder}{path to folder}

\item{width}{plot width}

\item{height}{plot height}
}
\description{
save plot lists
}
