% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPClusterPlot3DTree.R
\name{SNPClusterPlot3DTree}
\alias{SNPClusterPlot3DTree}
\title{plots cluster in 3D tree format}
\usage{
SNPClusterPlot3DTree(Cluster, angle = 60)
}
\arguments{
\item{Cluster}{object of HCPC \code{\link{SNPCluster}}}

\item{angle}{angle of tree}
}
\value{
Tree plot
}
\description{
plots cluster in 3D tree format
}
