% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPClusterPlotPCAMap.R
\name{SNPClusterPlotPCAMap}
\alias{SNPClusterPlotPCAMap}
\title{Plot SNP Cluster as a PCA Map}
\usage{
SNPClusterPlotPCAMap(Cluster)
}
\arguments{
\item{Cluster}{Cluster object of HCPC \code{\link{SNPCluster}}}
}
\value{
plot
}
\description{
Plot SNP Cluster as a PCA Map
}
